+++ example/read.js
@@ -1,11 +1,11 @@
 /* eslint-disable no-console */
-import fileJson from '../src'
+import bosom from '../src'
 import { resolve } from 'path'
 
 (async () => {
   try {
     const p = resolve(__dirname, 'example.json')
-    const res = await fileJson(p)
+    const res = await bosom(p)
     console.log(res)
   } catch (err) {
     console.error(err)