+++ test/spec/index.js
@@ -1,7 +1,7 @@
 import { equal, deepEqual } from 'zoroaster/assert'
 import SnapshotContext from 'snapshot-context'
 import Context from '../context'
-import fileJson from '../../src'
+import bosom from '../../src'
 
 /** @type {Object.<string, (c: Context, s: SnapshotContext)>} */
 const T = {
@@ -10,20 +10,20 @@ const T = {
     SnapshotContext,
   ],
   'is a function'() {
-    equal(typeof fileJson, 'function')
+    equal(typeof bosom, 'function')
   },
   async 'reads json from the file'({ data, path }) {
-    const json = await fileJson(path)
+    const json = await bosom(path)
     deepEqual(json, data)
   },
   async 'writes json data'({ data, tempPath }) {
-    await fileJson(tempPath, data)
+    await bosom(tempPath, data)
     const r = require(tempPath)
     deepEqual(r, data)
   },
   async 'writes json data with spaces'({ data, tempPath, SNAPSHOT_DIR, readTemp }, { setDir, test }) {
     setDir(SNAPSHOT_DIR)
-    await fileJson(tempPath, data, {
+    await bosom(tempPath, data, {
       space: 2,
     })
     const actual = await readTemp()